W160d: mt_evtMisc 4 window clears (003897f0 1308->828, 003890e0, 0038bc70, 0038ab00) + mdlManager shrinks; -1072 bytes, excess 5320->4360; __int128 aggregate-copy idiom, int/float round-trip fixes, short->int accumulator widening, long->int narrowing

+++ src/Event/mt_evtMisc.c
@@ -468,7 +468,7 @@ long FUN_00361ca0(int param_1,u64 param_2);
 long FUN_00361dd0(int param_1,u64 param_2,u16 param_3,u32 param_4);
 long FUN_00374af0(u64 param_1,u64 param_2,u64 param_3);
 int FUN_00386ae0(int param_1,int param_2);
-long FUN_0038ab00(int param_1,u64 param_2);
+int FUN_0038ab00(int param_1,u64 param_2);
 short FUN_0036f640(int param_1);
 u16 * FUN_00362290(int param_1);
 u16 * FUN_00362310(int param_1);
@@ -3639,7 +3639,7 @@ int FUN_0038a920(int param_1,u16 *param_2,int param_3)
 // FUN_0038AB00 NONMATCHING
 
 
-long FUN_0038ab00(int param_1,u64 param_2)
+int FUN_0038ab00(int param_1,u64 param_2)
 
 
 
@@ -3655,9 +3655,9 @@ long FUN_0038ab00(int param_1,u64 param_2)
 
   int *piVar5;
 
-  long lVar6;
+  int lVar6;
 
-  long lVar7;
+  int lVar7;
 
   int *piVar8;
